Amended controllers and test files to match new controller structure
