--- src/main/controllers/end_date/content.ts
@@ -1,14 +1,0 @@
-import { FormContent } from '../../definitions/form';
-import { AnyRecord } from '../../definitions/util-types';
-
-export const endDateContent: FormContent = {
-  fields: {},
-  submit: {
-    text: (l: AnyRecord): string => l.submit,
-    classes: 'govuk-!-margin-right-2',
-  },
-  saveForLater: {
-    text: (l: AnyRecord): string => l.saveForLater,
-    classes: 'govuk-button--secondary',
-  },
-};
